refactor(nox): split the noxfile into a support package
noxfile.py had grown to 1,745 lines of constants, runner plumbing, lane
implementations, and the verification graph around 22 sessions -- the
same shape that tripped the strict Sonar gate on the governance
checkers, waiting to fire on the next edit.

The sessions (and only the sessions) stay in noxfile.py (~400 lines);
everything else moves to tools/nox_support/: config (paths, limits, env
names), runner (SessionReporter, command execution, pytest/coverage
plumbing), policy_lanes (hygiene/policy/contracts/proof/lint),
test_lanes (tests/compatibility/fuzz/integration/docs/OSV), and graph
(the parallel verification graph and change-selected verification). A
commented facade import block keeps the helper surface reachable
through the noxfile module for the repo-policy test suite.

Test updates: the noxfile-introspecting tests patch globals where the
split modules actually resolve them via a new _patch_nox_globals helper
(patching a noxfile attribute alone can no longer reach a helper's
binding), and the three source-scan pins (positioning, identity
cutover, GIL assertions) point at the lane modules that now carry the
scanned text.

Verification: all 22 sessions register (nox -l); the tests, hygiene,
and policy lanes run green end-to-end through the split; the full
hermetic suite passes with the coverage gate; ruff format and lint
clean; check_repo_policy pass.
