+++ tools/nox_support/__init__.py
@@ -1,0 +1,1 @@
+"""Support package for the repository noxfile (constants, runner, lanes)."""
